scripts/requirements.txt
google-analytics-data==0.18.18
google-auth==2.35.0
